feat: enhancements and domain enrichments (#22)

+++ src/main/java/com/devt/metrics/domain/models/metrics/RepositoryMetric.java
@@ -3,9 +3,9 @@ package com.devt.metrics.domain.models.metrics;
 import java.util.List;
 
 public record RepositoryMetric(String name,
-                               int totalPRs,
-                               int totalReleases,
-                               int totalPipelines,
+                               long totalPRs,
+                               long totalReleases,
+                               long totalPipelines,
                                List<ReleaseMetric> releases,
                                List<PipelineMetric> pipelines,
                                PullRequestMetric pullRequests) {
